feat: add notification when absen masuk, absen pulang, cuti, sakit, and izin successfully submitted
This commit adds a notification feature to the app that will alert the user when their absen masuk, absen pulang, cuti, sakit, or izin has been successfully submitted.

This feature will improve the user experience by providing immediate feedback to the user and reducing the need to manually check the status of their submission.