feat: implement dayjs for date and time in izn screen
Refactor date and time field to use dayjs

This commit refactors the date and time fields in the izin screen to use dayjs instead of moment.js. Dayjs is a lightweight library for parsing, validating, and formatting dates and times. It is much smaller and faster than moment.js, and it has a simpler API.

This change will improve the performance and reduce the bundle size of the izin screen.