style: Implement global styles
Introduce a new 'styles' folder in the src directory.
Move common styles from individual screen component files to a central location in the 'styles' folder.
Create a 'styles.tsx' file to house the global styles.
Import and use the global styles in the screen component files.
This change allows for easier maintenance and consistency of styles across the application.