feat(LocationPicker): Add onDragMarker callback
This commit adds an onDragMarker callback to the LocationPicker component. The callback is called when the user drags the marker on the map, and it provides the new latitude and longitude of the marker.

This allows the developer to track the user's location as they drag the marker, and to update the state of the application accordingly.