refactor(hooks): rename useCurrentLocation hook to useLocation
This commit renames the useCurrentLocation hook to useLocation to better reflect its purpose. The hook now provides the current location of the user, regardless of whether they are currently clocked in or out. This change will make it easier for other developers to use the hook in their own applications.