refactor(Izin): refactor useLocation hook
This commit refactors the useLocation hook to use the new useLocation hook from the @src/hooks package.

The old useCurrentLocation hook was using a private API that is no longer available in the latest version of react-native-permissions. The new useLocation hook uses a public API that is supported in the latest version of react-native-permissions.